Restructure components folder
