+++ apps/admin/src/app/page.tsx
@@ -3,7 +3,7 @@
 import dynamic from 'next/dynamic'
 import { SessionProvider } from 'next-auth/react'
 
-const Admin = dynamic(() => import('./_components/Admin').then((mod) => mod.Admin), { ssr: false })
+const Admin = dynamic(() => import('../components/admin/Admin').then((mod) => mod.Admin), { ssr: false })
 
 const Home = () => (
   <main>
